move _set_rdescript to grammar manager
make descriptions visible for all rules
